servers/robot_controller_backend/requirements.txt
# Core dependencies
websockets>=11.0.3
asyncio-mqtt>=0.13.0
PyYAML>=6.0  # Configuration file parsing

# API Framework (for gateway_api.py)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
psutil>=5.9.0  # System metrics for performance endpoints

# Hardware control libraries
RPi.GPIO>=0.7.1
rpi-ws281x>=5.0.0  # NeoPixel/WS2812 LED strip control
adafruit-circuitpython-neopixel>=6.4.1
adafruit-circuitpython-dht>=3.4.0
adafruit-circuitpython-hcsr04>=3.3.0
adafruit-circuitpython-tsl2561>=3.2.0

# Camera libraries
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
picamera2>=0.3.12
Pillow>=10.0.0

# ROS2 Python bindings (for native ROS2 integration)
# Note: rclpy is typically installed with ROS2, but we list it here for reference
# Install via: sudo apt install ros-rolling-rclpy

# Sensor libraries
w1thermsensor>=2.0.0
adafruit-circuitpython-mpu6050>=3.2.0
adafruit-circuitpython-bno055>=3.2.0

# Serial communication
pyserial>=3.5

# Data processing
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning libraries
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
opencv-contrib-python>=4.8.0

# AI and Computer Vision
face-recognition>=1.3.0
dlib>=19.24.0
imutils>=0.5.4

# Logging and monitoring
structlog>=23.1.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0